select location and show weather

+++ app/src/main/java/android/podonin/com/perfectweather/model/DayShort.kt
@@ -8,29 +8,29 @@ data class DayShort(
     @SerializedName("condition")
     val condition: String,
     @SerializedName("feels_like")
-    val feelsLike: Int,
+    val feelsLike: Double,
     @SerializedName("humidity")
-    val humidity: Int,
+    val humidity: Double,
     @SerializedName("icon")
     val icon: String,
     @SerializedName("prec_mm")
-    val precMm: Int,
+    val precMm: Double,
     @SerializedName("prec_prob")
-    val precProb: Int,
+    val precProb: Double,
     @SerializedName("pressure_mm")
-    val pressureMm: Int,
+    val pressureMm: Double,
     @SerializedName("pressure_pa")
-    val pressurePa: Int,
+    val pressurePa: Double,
     @SerializedName("soil_moisture")
     val soilMoisture: Double,
     @SerializedName("soil_temp")
-    val soilTemp: Int,
+    val soilTemp: Double,
     @SerializedName("temp")
-    val temp: Int,
+    val temp: Double,
     @SerializedName("temp_min")
-    val tempMin: Int,
+    val tempMin: Double,
     @SerializedName("temp_water")
-    val tempWater: Int,
+    val tempWater: Double,
     @SerializedName("wind_dir")
     val windDir: String,
     @SerializedName("wind_gust")
