+++ app/src/main/java/android/podonin/com/perfectweather/model/Fact.kt
@@ -42,5 +42,5 @@ data class Fact(
     @SerializedName("wind_gust")
     val windGust: Double,
     @SerializedName("wind_speed")
-    val windSpeed: Int
+    val windSpeed: Double
 )
