+++ app/src/main/java/android/podonin/com/perfectweather/model/WetherResponse.kt
@@ -10,7 +10,7 @@ data class WetherResponse(
     @SerializedName("info")
     val info: Info,
     @SerializedName("now")
-    val now: Int,
+    val now: Double,
     @SerializedName("now_dt")
     val nowDt: String
 )
